Refactor ErrorPage component to have buttons to sign in and to move to home page
