+++ src/domain/auth/logoutPage/LogoutPage.tsx
@@ -1,15 +1,12 @@
-import { IconArrowLeft } from 'hds-react';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
 
-import Button from '../../../common/components/button/Button';
-import ErrorTemplate from '../../../common/components/errorTemplate/ErrorTemplate';
+import ErrorPage from '../../../common/components/errorPage/ErrorPage';
 import LoadingSpinner from '../../../common/components/loadingSpinner/LoadingSpinner';
 import { ROUTES } from '../../../constants';
 import useLocale from '../../../hooks/useLocale';
 import getValue from '../../../utils/getValue';
-import MainContent from '../../app/layout/mainContent/MainContent';
 import PageWrapper from '../../app/layout/pageWrapper/PageWrapper';
 import { useAuth } from '../hooks/useAuth';
 import styles from './logoutPage.module.scss';
@@ -19,15 +16,6 @@ const LogoutPage: React.FC = () => {
   const { t } = useTranslation();
   const locale = useLocale();
   const navigate = useNavigate();
-  const { signIn } = useAuth();
-
-  const goToHome = () => {
-    navigate(`/${locale}${ROUTES.HOME}`);
-  };
-
-  const handleSignIn = () => {
-    signIn(`/${locale}${ROUTES.HOME}`);
-  };
 
   React.useEffect(() => {
     if (isAuthenticated) {
@@ -44,32 +32,11 @@ const LogoutPage: React.FC = () => {
 
   return (
     <PageWrapper title={getValue(t('logoutPage.pageTitle'), '')}>
-      <MainContent>
-        <ErrorTemplate
-          buttons={
-            <div className={styles.buttons}>
-              <Button
-                fullWidth={true}
-                iconLeft={<IconArrowLeft />}
-                onClick={goToHome}
-                type="button"
-                variant="secondary"
-              >
-                {t('logoutPage.buttonBackToHome')}
-              </Button>
-              <Button
-                fullWidth={true}
-                onClick={handleSignIn}
-                type="button"
-                variant="primary"
-              >
-                {t('common.signIn')}
-              </Button>
-            </div>
-          }
-          text={t('logoutPage.text')}
-        />
-      </MainContent>
+      <ErrorPage
+        homePageButtonText={t('logoutPage.buttonBackToHome')}
+        signInPath={`/${locale}${ROUTES.HOME}`}
+        text={t('logoutPage.text')}
+      />
     </PageWrapper>
   );
 };
