+++ src/domain/auth/logoutPage/logoutPage.module.scss
@@ -1,13 +1,3 @@
-.buttons {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: var(--spacing-m);
-
-  button {
-    white-space: nowrap;
-  }
-}
-
 .loadingSpinner {
   margin-top: var(--spacing-5-xl);
 }
