--- src/domain/notFound/notFound.module.scss
@@ -1,5 +1,0 @@
-.buttons {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: var(--spacing-m);
-}
