--- src/domain/notSigned/notSigned.module.scss
@@ -1,5 +1,0 @@
-.buttons {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: var(--spacing-m);
-}
